pyproject.toml
[project]
name = "doric-dataset"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.0.1",
    "openai>=1.0.0",
    "modal>=1.2.2",
    "tinker>=0.3.0",
    "pyright>=1.1.407",
    "transformers>=4.57.1",
    "instructor>=1.13.0",
]
